+++ subsys/net/ip/l2/ieee802154/ieee802154.c
@@ -28,7 +28,7 @@
 #include <net/ieee802154_radio.h>
 
 #include "ieee802154_frame.h"
-#include "ieee802154_mgmt.h"
+#include "ieee802154_mgmt_priv.h"
 #include "ieee802154_security.h"
 #include "ieee802154_utils.h"
